feat(matcher): implement step3 sharding and risk
Execute Step 3 from docs/implementation-steps/03-sharding-and-risk.md as authoritative scope.
- Add deterministic shard router and single-thread shard runtime
- Add market/outcome command metadata and multi-book management
- Add risk checks before reservation with reject counters
- Add migration path with replay hash verification and admin endpoints
- Add routing stability, migration, risk, and multi-market determinism tests
- Keep workspace quality gates green (fmt/test/clippy/build)

+++ ledger-service/src/logic.rs
@@ -85,7 +85,7 @@ mod tests {
                 let mut reserved = reserve;
 
                 // reserve
-                let reserve_amt = (available_start % 4) as i64;
+                let reserve_amt = available_start % 4;
                 if available >= reserve_amt {
                     available -= reserve_amt;
                     reserved += reserve_amt;
@@ -93,7 +93,7 @@ mod tests {
                 assert!(available >= 0 && reserved >= 0);
 
                 // spend from reserved first, then available
-                let spend = (available_start % 5) as i64;
+                let spend = available_start % 5;
                 let total = available + reserved;
                 if total >= spend {
                     let consume_reserved = reserved.min(spend);
